# Add JsonNumberHandling(Attribute) & support for (de)serializing numbers from/to string