--- a/src/libraries/System.Text.Json/src/System/Text/Json/JsonConstants.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/JsonConstants.cs
@@ -1,8 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Text.Json.Serialization;
-
 namespace System.Text.Json
 {
     internal static class JsonConstants
@@ -85,6 +83,7 @@ namespace System.Text.Json
             (DateTimeParseNumFractionDigits - DateTimeNumFractionDigits)); // Like StandardFormat 'O' for DateTimeOffset, but allowing 9 additional (up to 16) fraction digits.
         public const int MinimumDateTimeParseLength = 10; // YYYY-MM-DD
         public const int MaximumEscapedDateTimeOffsetParseLength = MaxExpansionFactorWhileEscaping * MaximumDateTimeOffsetParseLength;
+        public const int NegativeInfinityLiteralConstantLength = 9; // Character count for -Infinity is 9
 
         internal const char ScientificNotationFormat = 'e';
 
@@ -107,8 +106,5 @@ namespace System.Text.Json
 
         // The maximum number of parameters a constructor can have where it can be supported.
         public const int MaxParameterCount = 64;
-
-        public const JsonNumberHandling ReadNumberOrFloatingConstantFromString = JsonNumberHandling.AllowReadingFromString | JsonNumberHandling.AllowNamedFloatingPointLiterals;
-        public const JsonNumberHandling WriteNumberOrFloatingConstantAsString = JsonNumberHandling.WriteAsString | JsonNumberHandling.AllowNamedFloatingPointLiterals;
     }
 }