--- a/src/libraries/System.Text.Json/src/System/Text/Json/Reader/JsonReaderHelper.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Reader/JsonReaderHelper.cs
@@ -321,5 +321,18 @@ namespace System.Text.Json
             value = default;
             return false;
         }
+
+        public static char GetFloatingPointStandardParseFormat(ReadOnlySpan<byte> span)
+        {
+            foreach (byte token in span)
+            {
+                if (token == 'E' || token == 'e')
+                {
+                    return JsonConstants.ScientificNotationFormat;
+                }
+            }
+
+            return default;
+        }
     }
 }