--- a/src/libraries/System.Text.Json/src/System/Text/Json/Reader/JsonReaderHelper.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Reader/JsonReaderHelper.cs
@@ -341,7 +341,7 @@ namespace System.Text.Json
         {
             if (span.Length == 3)
             {
-                if (ValueIsNan(span))
+                if (span.SequenceEqual(JsonConstants.NaNValue))
                 {
                     value = float.NaN;
                     return true;
@@ -349,15 +349,15 @@ namespace System.Text.Json
             }
             else if (span.Length == 8)
             {
-                if (ValueIsPositiveInfinity(span))
+                if (span.SequenceEqual(JsonConstants.PositiveInfinityValue))
                 {
                     value = float.PositiveInfinity;
                     return true;
                 }
             }
-            else if (span.Length == JsonConstants.NegativeInfinityLiteralConstantLength)
+            else if (span.Length == 9)
             {
-                if (ValueIsNegativeInfinity(span))
+                if (span.SequenceEqual(JsonConstants.NegativeInfinityValue))
                 {
                     value = float.NegativeInfinity;
                     return true;
@@ -372,7 +372,7 @@ namespace System.Text.Json
         {
             if (span.Length == 3)
             {
-                if (ValueIsNan(span))
+                if (span.SequenceEqual(JsonConstants.NaNValue))
                 {
                     value = double.NaN;
                     return true;
@@ -380,15 +380,15 @@ namespace System.Text.Json
             }
             else if (span.Length == 8)
             {
-                if (ValueIsPositiveInfinity(span))
+                if (span.SequenceEqual(JsonConstants.PositiveInfinityValue))
                 {
                     value = double.PositiveInfinity;
                     return true;
                 }
             }
-            else if (span.Length == JsonConstants.NegativeInfinityLiteralConstantLength)
+            else if (span.Length == 9)
             {
-                if (ValueIsNegativeInfinity(span))
+                if (span.SequenceEqual(JsonConstants.NegativeInfinityValue))
                 {
                     value = double.NegativeInfinity;
                     return true;
@@ -397,39 +397,6 @@ namespace System.Text.Json
 
             value = 0;
             return false;
-        }
-
-        private static bool ValueIsNan(ReadOnlySpan<byte> span)
-        {
-            Debug.Assert(span.Length == 3);
-            return span[0] == (byte)'N' && span[1] == (byte)'a' && span[2] == (byte)'N';
-        }
-
-        private static bool ValueIsPositiveInfinity(ReadOnlySpan<byte> span)
-        {
-            Debug.Assert(span.Length == 8);
-            return span[0] == (byte)'I' &&
-                span[1] == (byte)'n' &&
-                span[2] == (byte)'f' &&
-                span[3] == (byte)'i' &&
-                span[4] == (byte)'n' &&
-                span[5] == (byte)'i' &&
-                span[6] == (byte)'t' &&
-                span[7] == (byte)'y';
-        }
-
-        private static bool ValueIsNegativeInfinity(ReadOnlySpan<byte> span)
-        {
-            Debug.Assert(span.Length == JsonConstants.NegativeInfinityLiteralConstantLength);
-            return span[0] == (byte)'-' &&
-                span[1] == (byte)'I' &&
-                span[2] == (byte)'n' &&
-                span[3] == (byte)'f' &&
-                span[4] == (byte)'i' &&
-                span[5] == (byte)'n' &&
-                span[6] == (byte)'i' &&
-                span[7] == (byte)'t' &&
-                span[8] == (byte)'y';
         }
     }
 }