Fixes https://github.com/dotnet/runtime/issues/30255.