@layomia 2020-07-21
Barring benchmark noise, there are no noticeable perf regressions caused by this PR:

```
dotnet run --base D:\benchmarks\numbers\master --diff D:\benchmarks\numbers\feature --threshold 2%
summary:
better: 4, geomean: 1.039
worse: 4, geomean: 1.034
total diff: 8
```

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Serialization.Tests.ReadJson<Int32>.DeserializeFromUtf8Bytes    |      1.04 |            81.42 |            85.00 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Hashtable>.SerializeToUtf8Bytes   |      1.04 |         15760.37 |         16318.96 |         |
| System.Text.Json.Serialization.Tests.WriteJson<LoginViewModel>.SerializeToUtf8By |      1.03 |           318.62 |           328.68 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Int32>.SerializeToUtf8Bytes       |      1.03 |           120.47 |           123.72 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Serialization.Tests.WriteJson<BinaryData>.SerializeToUtf8Bytes  |      1.04 |           416.34 |           400.33 |         |
| System.Text.Json.Serialization.Tests.WriteJson<MyEventsListerViewModel>.Serializ |      1.04 |        462715.53 |        445168.06 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Location>.SerializeToUtf8Bytes    |      1.04 |           808.07 |           778.40 |         |
| System.Text.Json.Serialization.Tests.ReadJson<SimpleStructWithProperties>.Deseri |      1.04 |           304.35 |           293.22 |         |

[feature.zip](https://github.com/dotnet/runtime/files/4950750/feature.zip)
[master.zip](https://github.com/dotnet/runtime/files/4950752/master.zip)